requirements.txt
drf-yasg == 1.21.10
Django == 4.2.23
djangorestframework == 3.16.0
djangorestframework_simplejwt == 5.5.0
pytest-django == 4.11.1
